requirements.txt
mkdocs>=1.6
mkdocs-material>=9.5

# Geospatial data harmonization
rasterio>=1.3
numpy>=1.21
matplotlib>=3.5
shapely>=2.0
geopandas>=0.14
folium>=0.14
branca>=0.6

# Configuration support for examples
pyyaml>=6.0

# STAC client for cloud-native data discovery (Sentinel-2, Landsat, NAIP, etc.)
pystac-client>=0.7

# Climate / scientific array formats (NetCDF, OPeNDAP, Zarr)
xarray>=2024.0
netCDF4>=1.6
zarr>=2.0

# Testing and catalog sync
pytest>=7.0
requests>=2.28
